requirements.txt
# Core
facenet-pytorch>=2.5.3
opencv-python-headless>=4.8
numpy>=1.24
pillow>=10.0
torch>=2.0
torchvision>=0.15
scipy>=1.11

# API
fastapi>=0.109
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.5
pydantic-settings>=2.1

# Database
aiosqlite>=0.19

# Config
pyyaml>=6.0
python-dotenv>=1.0

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.23
ruff>=0.1
pre-commit>=3.0
httpx>=0.26
